+++ src/Cake.ServiceOrchestration/IServiceManager.cs
@@ -53,9 +53,9 @@ namespace Cake.ServiceOrchestration
         void DeployService();
 
         /// <summary>
-        ///     Deploys all instances of the current service that match the given predicate.
+        ///     Deploys all instances of the current service that match the given filter.
         /// </summary>
         /// <remarks>Implementations may use varying logic for this process</remarks>
-        void DeployService(Func<IServiceInstance, bool> predicate);
+        void DeployService(IServiceFilter filter);
     }
 }