+++ ReleaseNotes.md
@@ -10,3 +10,4 @@
 # 0.3.0
 
 - Introduces tags for instances
+- Introduces support for instance filtering when deploying (using `IServiceFilter`)